feat(iot-sdk、iot-manager): 补充完善设备控制（未完成）

+++ pangu-iot/pangu-iot-manager/src/main/java/com/pangu/iot/manager/product/domain/ProductService.java
@@ -40,7 +40,7 @@ public class ProductService extends BaseEntity {
     /**
      * 执行方式 0-同步 1-异步
      */
-    private Long async;
+    private Integer async;
 
 //    @TableField(exist = false)
 //    private Long relationId;
